pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-data-langroid"
version = "0.1.0"
description = "Agent Data Langroid - Multi-agent knowledge management system built with Langroid framework"
authors = [{name = "Agent Data Team", email = "team@agentdata.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11,<3.13"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    # Core Langroid framework with extras
    "langroid>=0.58.0,<0.59",

    # Web framework and API
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",

    # Google Cloud services
    "google-cloud-storage>=2.10.0",
    "google-cloud-firestore>=2.13.1",
    "google-auth>=2.23.0",
    "google-api-python-client>=2.108.0",

    # Vector database
    "qdrant-client>=1.6.0",

    # AI and ML
    "openai>=1.3.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",

    # Data processing
    "pandas>=2.1.0",
    "pydantic>=2.4.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "rich>=13.5.0",
    "typer>=0.9.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",

    # Code quality
    "black>=23.9.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0",

    # Documentation
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.4.0",

    # Dependency management
    "pip-tools>=7.3.0",
]

test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "httpx>=0.25.0",  # For testing FastAPI endpoints
]

docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.4.0",
    "mkdocs-mermaid2-plugin>=0.6.0",
]

[project.urls]
Homepage = "https://github.com/your-org/agent-data-langroid"
Repository = "https://github.com/your-org/agent-data-langroid.git"
Issues = "https://github.com/your-org/agent-data-langroid/issues"
Documentation = "https://your-org.github.io/agent-data-langroid/"

[project.scripts]
agent-data = "agent_data.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["agent_data*"]
exclude = ["tests*"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=agent_data",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
testpaths = ["tests"]
filterwarnings = [
    "error",
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "e2e: marks tests as end-to-end tests",
    "langroid: marks tests that require Langroid",
    "mock: marks tests that use mocks",
    "qdrant: marks tests that require Qdrant",
    "shadow: marks tests for shadow traffic functionality",
]

[tool.coverage.run]
source = ["agent_data"]
omit = [
    "*/tests/*",
    "*/venv/*",
    "*/.venv/*",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.black]
line-length = 88
target-version = ["py311"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["agent_data"]
known_third_party = ["langroid"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "qdrant_client.*",
    "google.cloud.*",
    "langroid.*",
]
ignore_missing_imports = true

[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
